pyproject.toml
[project]
name = "tenderscraper"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "httpx>=0.27",
  "tenacity>=8.2",
  "pydantic>=2.6",
  "sqlmodel>=0.0.22",
  "python-dotenv>=1.0",
  "typer>=0.12",
  "pydantic-settings>=2.0",
  "playwright>=1.45",
]

[project.optional-dependencies]
dev = [
  "black>=24.0",
  "ruff>=0.6",
  "isort>=5.13",
  "mypy>=1.11",
  "pytest>=8.0",
]

[project.scripts]
tenderscraper = "tenderscraper.cli:app"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "B", "UP", "D"]
ignore = [
  "D100","D101","D102","D103","D104","D105","D107"
]
[tool.ruff.pydocstyle]
convention = "google"
[tool.ruff.format]
quote-style = "preserve"

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true